refactor: split kali-server request handling modules
