+++ cmd/kali-server/helpers.go
@@ -1,0 +1,73 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/found-cake/kali-mcp-go/internal/executor"
+	"github.com/found-cake/kali-mcp-go/internal/tools"
+	"github.com/found-cake/kali-mcp-go/pkg/dto"
+	"github.com/gofiber/fiber/v3"
+)
+
+func toolStatus(lookup func(string) bool) map[string]bool {
+	status := make(map[string]bool, len(exposedToolBinaries))
+	dirWordlistReady := tools.WordlistExists(tools.DefaultDirWordlistPath())
+	johnWordlistReady := tools.WordlistExists(tools.DefaultJohnWordlistPath())
+
+	for _, toolName := range exposedToolBinaries {
+		ready := lookup(toolName)
+		switch toolName {
+		case "gobuster", "dirb":
+			ready = ready && dirWordlistReady
+		case "john":
+			ready = ready && johnWordlistReady
+		}
+		status[toolName] = ready
+	}
+
+	return status
+}
+
+func allEssentialToolsAvailable(status map[string]bool) bool {
+	for _, toolName := range essentialToolBinaries {
+		if !status[toolName] {
+			return false
+		}
+	}
+	return true
+}
+
+func toAPIResult(r *executor.Result) dto.ToolResult {
+	return dto.ToolResult{
+		Stdout:         r.Stdout,
+		Stderr:         r.Stderr,
+		ReturnCode:     r.ReturnCode,
+		Success:        r.Success(),
+		TimedOut:       r.TimedOut,
+		PartialResults: r.TimedOut && (r.Stdout != "" || r.Stderr != ""),
+	}
+}
+
+func badRequest(c fiber.Ctx, msg string) error {
+	return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": msg})
+}
+
+func internalServerError(c fiber.Ctx, msg string) error {
+	return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": msg})
+}
+
+func bindJSON(c fiber.Ctx, dst any) error {
+	return c.Bind().Body(dst)
+}
+
+func parseRequest[T any](c fiber.Ctx, validate func(T) error) (T, error) {
+	var req T
+	bindErr := bindJSON(c, &req)
+	if bindErr != nil {
+		return req, fmt.Errorf("invalid request body")
+	}
+	if err := validate(req); err != nil {
+		return req, err
+	}
+	return req, nil
+}
